extracted sort data function
